Complete JS-Racer
This was a pretty fun challenge. I did my best to explore
concepts from class yesterday. I experimented heavily with
timed events as well as the key-binding listeners. I'm not
happy with the length of several of my functions, but I'm
uncertain of how to compress the logic further. I attempted
to refactor the advancePlayer functions into one function that
took a passed in id as an argument but was unable to get it to
work as desired. Overall, I'm pleased with the progress I've
made with jQuery and JS.